Corrected some warnings, corrected testscript file.

+++ Twin64-Simulator/T64-SimWinClasses.cpp
@@ -114,7 +114,7 @@ int calculateStrLen( const char *s, int tabWidth  ) {
         s++;
     }
 
-    return col;
+    return ( static_cast<int> ( col ));
 }
 
 //----------------------------------------------------------------------------------------
